Revert "Merge remote-tracking branch 'origin/fix/issue-271-float32-value-bus' into integration/0.3.0-entity"
This reverts commit 4dd6fb74d958a0e29d76a9ee1815d7e08ab8481d, reversing
changes made to 47c2f49b753ad9e2235c4a34a205887448c38871.

+++ internal/agent/probes/host/wifiSignalStrengthProbe.go
@@ -196,7 +196,7 @@ func (m *wifiSignalStrengthProbe) collectWindows() ([]data_store.DataPoint, erro
 		{
 			Name:      "wifi_signal_strength",
 			Timestamp: time.Now(),
-			Value:     float64(signalStrength),
+			Value:     float32(signalStrength),
 			Tags:      wifiTags,
 		},
 	}, nil
@@ -242,7 +242,7 @@ func (m *wifiSignalStrengthProbe) collectLinux() ([]data_store.DataPoint, error)
 				dataPoints = append(dataPoints, data_store.DataPoint{
 					Name:      "wifi_signal_strength",
 					Timestamp: timestamp,
-					Value:     float64(signalStrength),
+					Value:     float32(signalStrength),
 					Tags:      wifiTags,
 				})
 			}
@@ -254,7 +254,7 @@ func (m *wifiSignalStrengthProbe) collectLinux() ([]data_store.DataPoint, error)
 					quality, err := strconv.Atoi(qualityParts[0])
 					maxQuality, err2 := strconv.Atoi(qualityParts[1])
 					if err == nil && err2 == nil && maxQuality > 0 {
-						qualityPercent := float64(quality) / float64(maxQuality) * 100
+						qualityPercent := float32(quality) / float32(maxQuality) * 100
 						dataPoints = append(dataPoints, data_store.DataPoint{
 							Name:      "wifi_quality",
 							Timestamp: timestamp,
